+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- ⚠️ **BREAKING CHANGE** ⚠️
+
+  The `RegularSchedule` in the `beat` module has been renamed to `DeltaSchedule` to
+  be more coherent with Python Celery terminology, where it is sometimes called *timedelta*.
+
 ### Added
 
 - 🚀🚀 Redis broker support 🚀🚀